backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
